refactor: isolate PDF converters
