+++ haystack/nodes/file_converter/__init__.py
@@ -8,6 +8,7 @@ from haystack.nodes.file_converter.tika import TikaConverter, TikaXHTMLParser
 from haystack.nodes.file_converter.txt import TextConverter
 from haystack.nodes.file_converter.azure import AzureConverter
 from haystack.nodes.file_converter.parsr import ParsrConverter
+from haystack.nodes.file_converter.pdf import PDFToTextConverter
 
 MarkdownConverter = safe_import(
     "haystack.nodes.file_converter.markdown", "MarkdownConverter", "preprocessing"
@@ -15,9 +16,6 @@ MarkdownConverter = safe_import(
 ImageToTextConverter = safe_import(
     "haystack.nodes.file_converter.image", "ImageToTextConverter", "ocr"
 )  # Has optional dependencies
-PDFToTextConverter = safe_import(
-    "haystack.nodes.file_converter.pdf", "PDFToTextConverter", "ocr"
-)  # Has optional dependencies
 PDFToTextOCRConverter = safe_import(
-    "haystack.nodes.file_converter.pdf", "PDFToTextOCRConverter", "ocr"
+    "haystack.nodes.file_converter.pdf_ocr", "PDFToTextOCRConverter", "ocr"
 )  # Has optional dependencies
